chore: refactor configuration page and add missing docstrings
